+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2020-02-26 - Fix a behat test and added some details to the plugin settings page after the 'Extend session' popup was introduced upstream in MDL-34498.
 * 2020-02-26 - Ship the simple Ajax endpoint sessionkeepalive_ajax.php with this plugin as it was removed upstream in MDL-34498.
 * 2020-02-26 - Prepare compatibility for Moodle 3.8.